# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hospital Management API",
    "description": "CSE 341 Project 2 API",
    "version": "1.0.0"
  },
  "host": "hospital-management-api-jp2e.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "GitHubOAuth": {
      "type": "oauth2",
      "authorizationUrl": "https://github.com/login/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "user:email": "Access user email"
      }
    }
  },
  "paths": {
...
    "/auth/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "View logged-in user",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "GitHubOAuth": []
          }
        ]
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "GitHubOAuth": []
          }
        ]
      }
    },
...
    "/profile": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "View logged-in user",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "GitHubOAuth": []
          }
        ]
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "GitHubOAuth": []
          }
        ]
      }
    },
...
  }
}